+++ src/zigbee2mqtt.ts
@@ -317,6 +317,12 @@ export class Zigbee2MQTT extends EventEmitter {
         this.log.error('Error creating data directory:', error);
       }
     }
+    try {
+      const filePath = path.join(this.mqttDataPath, 'bridge-payloads.txt');
+      fs.unlinkSync(filePath);
+    } catch (error) {
+      this.log.debug(`Error deleting bridge-payloads.txt: ${error}`);
+    }
   }
 
   // Get the URL for connect
@@ -906,6 +912,7 @@ export class Zigbee2MQTT extends EventEmitter {
       // Do nothing
     } else if (service === undefined) {
       // this.log.debug(`classZigbee2MQTT=>emitting message for device ${dn}${entity}${rs} payload ${pl}${payload}${rs}`);
+      this.emit('message', entity, data);
       this.emit('MESSAGE-' + entity, data);
     } else {
       // MQTT output attribute type